refactor(launcher): split runtime ownership boundaries
refactor: extract runtime execution, integration, route, setup, transaction, supervisor control, startup, stop, tunnel, and helper responsibilities into modules capped below 500 lines.

fix: merge the 3.0.0 isolated DEV harness contract into the split architecture, including profile validation, tunnel-only supervision, diagnostics, setup environment isolation, cancellation, forced shutdown, stale-version recovery, and the 24-hour tunnel lifetime.

test: establish a 70-pass/13-fail refactor RED, restore all 83 focused runtime contracts, pass the complete Launcher suite with 185 passes and one intentional platform skip, then pass Launcher TypeScript and renderer production build.
