+++ launcher/electron/runtime-execution.cjs
@@ -1,0 +1,166 @@
+const path = require("node:path");
+const fs = require("node:fs");
+const os = require("node:os");
+const { randomBytes } = require("node:crypto");
+const { spawn } = require("node:child_process");
+const { writePrivateFileAtomic } = require("./atomic-file.cjs");
+const { readJsonFile } = require("./json-file.cjs");
+const { connectorNameForDevSetup, connectorNameForSetup, CURRENT_CONNECTOR_NAME, isLegacyConnectorName, requireCurrentRuntimeConnectorName, validateConnectorName } = require("./connector-identity.cjs");
+const { embeddedRuntimeInvocation, runtimeInvocation } = require("./runtime-command.cjs");
+const { redactText } = require("./logging.cjs");
+const { DETACH_OWNED_CHILD, terminateOwnedProcessTree } = require("./process-tree.cjs");
+const helpers = require("./runtime-helpers.cjs");
+const { MAX_CAPTURE_BYTES, MAX_RUNTIME_LOG_LINE_CHARS, CORE_SETUP_TIMEOUT_MS, MCP_SETUP_TIMEOUT_MS, UNINSTALL_TIMEOUT_MS, MAX_CHECKPOINT_FILE_BYTES, collect, resolveUserPath, captureRegularFile, restoreRegularFile, regularFileChanged, parseBridgeRouteResult } = helpers;
+
+module.exports = {
+  async run(name, args, options = {}) {
+    if (this.active) throw new Error(`Another launcher operation is active: ${this.active}`);
+    if (this.activeChild
+      && this.activeChild.exitCode === null
+      && this.activeChild.signalCode === null) {
+      throw new Error("A previous launcher operation process is still running");
+    }
+    this.activeChild = null;
+    if (this.lifecycleOperation && this.lifecycleOperation !== name) {
+      throw new Error(`Another launcher operation is active: ${this.lifecycleOperation}`);
+    }
+    this.active = name;
+    this.publishOperation?.({ name, status: "running", message: options.message || name });
+    this.logger.info("runtime.operation_started", { name, args: args.map((arg) => /key|token/i.test(arg) ? "[redacted]" : arg) });
+    try {
+      const invocation = options.embedded
+        ? embeddedRuntimeInvocation({ app: this.app, sourceRoot: this.sourceRoot, args })
+        : this.command(args);
+      const result = await new Promise((resolve, reject) => {
+        const environment = options.environment
+          ? { ...options.environment }
+          : { ...process.env };
+        Object.assign(environment, {
+          CODEX_CHATGPT_WEB_BROWSER_HOST_DESCRIPTOR: this.browserDescriptorPath,
+          ...(options.env || {}),
+        });
+        const child = spawn(invocation.executable, invocation.args, {
+          cwd: invocation.cwd,
+          detached: DETACH_OWNED_CHILD,
+          env: environment,
+          stdio: ["ignore", "pipe", "pipe"],
+          windowsHide: true,
+        });
+        this.activeChild = child;
+        const stdout = [];
+        const stderr = [];
+        const pipeErrors = [];
+        const recordPipeError = (stream) => (error) => {
+          pipeErrors.push(`${name} ${stream} pipe failed: ${error instanceof Error ? error.message : String(error)}`);
+        };
+        collect(child.stdout, stdout, (line) => {
+          this.logger.info("runtime.stdout", { operation: name, line });
+          this.publishOperation?.({ name, status: "running", message: redactText(line) });
+        }, recordPipeError("stdout"));
+        collect(child.stderr, stderr, (line) => {
+          this.logger.warn("runtime.stderr", { operation: name, line });
+          this.publishOperation?.({ name, status: "running", message: redactText(line) });
+        }, recordPipeError("stderr"));
+        let settled = false;
+        let timedOut = null;
+        let terminationTimeout = null;
+        let forceTimeout = null;
+        const clearTimers = () => {
+          if (timeout) clearTimeout(timeout);
+          if (terminationTimeout) clearTimeout(terminationTimeout);
+          if (forceTimeout) clearTimeout(forceTimeout);
+        };
+        const timeout = options.timeoutMs
+          ? setTimeout(() => {
+              if (settled) return;
+              timedOut = new Error(`${name} timed out after ${options.timeoutMs}ms`);
+              try {
+                terminateOwnedProcessTree(child);
+              } catch (error) {
+                settled = true;
+                clearTimers();
+                reject(new Error(
+                  `${timedOut.message}; child process tree termination failed: ${error instanceof Error ? error.message : String(error)}`,
+                ));
+                return;
+              }
+              terminationTimeout = setTimeout(() => {
+                if (settled) return;
+                try {
+                  terminateOwnedProcessTree(child, "SIGKILL");
+                } catch (error) {
+                  settled = true;
+                  clearTimers();
+                  reject(new Error(
+                    `${timedOut.message}; forced child process tree termination failed: ${error instanceof Error ? error.message : String(error)}`,
+                  ));
+                  return;
+                }
+                forceTimeout = setTimeout(() => {
+                  if (settled) return;
+                  settled = true;
+                  clearTimers();
+                  reject(new Error(`${timedOut.message}; the child process did not exit after forced termination`));
+                }, 2_000);
+              }, 5_000);
+            }, options.timeoutMs)
+          : null;
+        child.once("error", (error) => {
+          const childStillRunning = Number.isInteger(child.pid)
+            && child.exitCode === null
+            && child.signalCode === null;
+          if (this.activeChild === child && !childStillRunning) this.activeChild = null;
+          if (settled) return;
+          settled = true;
+          clearTimers();
+          reject(timedOut
+            ? new Error(`${timedOut.message}; termination failed: ${error.message}`)
+            : error);
+        });
+        child.once("exit", (code, signal) => {
+          if (this.activeChild === child) this.activeChild = null;
+          if (settled) return;
+          settled = true;
+          clearTimers();
+          if (timedOut) {
+            try {
+              terminateOwnedProcessTree(child, "SIGKILL");
+              reject(timedOut);
+            } catch (error) {
+              reject(new Error(
+                `${timedOut.message}; final process-group cleanup failed: ${error instanceof Error ? error.message : String(error)}`,
+              ));
+            }
+            return;
+          }
+          if (pipeErrors.length > 0) {
+            reject(new Error(pipeErrors.join("; ")));
+            return;
+          }
+          resolve({
+            code: code ?? 1,
+            signal,
+            stdout: Buffer.concat(stdout).toString("utf8"),
+            stderr: Buffer.concat(stderr).toString("utf8"),
+          });
+        });
+      });
+      const acceptedExitCodes = options.acceptedExitCodes || [0];
+      if (!acceptedExitCodes.includes(result.code)) {
+        const detail = result.stderr.trim() || result.stdout.trim() || `exit ${result.code}`;
+        throw new Error(detail);
+      }
+      this.logger.info("runtime.operation_completed", { name });
+      this.publishOperation?.({ name, status: "completed", message: options.successMessage || "Completed" });
+      return result;
+    } catch (error) {
+      const message = redactText(error instanceof Error ? error.message : String(error));
+      this.logger.error("runtime.operation_failed", { name, message });
+      this.publishOperation?.({ name, status: "failed", message });
+      throw new Error(message);
+    } finally {
+      this.active = null;
+    }
+  },
+
+};
