+++ launcher/electron/runtime-helpers.cjs
@@ -1,0 +1,139 @@
+const path = require("node:path");
+const fs = require("node:fs");
+const os = require("node:os");
+const { randomBytes } = require("node:crypto");
+const { spawn } = require("node:child_process");
+const { writePrivateFileAtomic } = require("./atomic-file.cjs");
+const {
+  connectorNameForSetup,
+  CURRENT_CONNECTOR_NAME,
+  isLegacyConnectorName,
+  requireCurrentRuntimeConnectorName,
+  validateConnectorName,
+} = require("./connector-identity.cjs");
+const { embeddedRuntimeInvocation, runtimeInvocation } = require("./runtime-command.cjs");
+const { redactText } = require("./logging.cjs");
+const { DETACH_OWNED_CHILD, terminateOwnedProcessTree } = require("./process-tree.cjs");
+
+const MAX_CAPTURE_BYTES = 8 * 1024 * 1024;
+const MAX_RUNTIME_LOG_LINE_CHARS = 64 * 1024;
+const CORE_SETUP_TIMEOUT_MS = 5 * 60_000;
+const MCP_SETUP_TIMEOUT_MS = 10 * 60_000;
+const UNINSTALL_TIMEOUT_MS = 2 * 60_000;
+const MAX_CHECKPOINT_FILE_BYTES = 16 * 1024 * 1024;
+function collect(stream, chunks, onLine, onError) {
+  let buffered = "";
+  let bytes = 0;
+  stream.on("data", (chunk) => {
+    bytes += chunk.length;
+    if (bytes <= MAX_CAPTURE_BYTES) chunks.push(chunk);
+    buffered += chunk.toString("utf8");
+    for (;;) {
+      const newline = buffered.indexOf("\n");
+      if (newline < 0) break;
+      const line = buffered.slice(0, newline).trimEnd();
+      buffered = buffered.slice(newline + 1);
+      if (line) onLine(line);
+    }
+    if (buffered.length > MAX_RUNTIME_LOG_LINE_CHARS) {
+      onLine(`${buffered.slice(0, MAX_RUNTIME_LOG_LINE_CHARS)}…[truncated]`);
+      buffered = "";
+    }
+  });
+  stream.on("end", () => {
+    const line = buffered.trim();
+    if (line) onLine(line);
+  });
+  stream.on("error", (error) => onError?.(error));
+}
+
+function resolveUserPath(value) {
+  if (value === "~") return os.homedir();
+  if (value.startsWith("~/") || value.startsWith("~\\")) {
+    return path.resolve(os.homedir(), value.slice(2));
+  }
+  return path.resolve(value);
+}
+
+function captureRegularFile(filePath) {
+  let stat;
+  try {
+    stat = fs.lstatSync(filePath);
+  } catch (error) {
+    if (error?.code === "ENOENT") return { path: filePath, exists: false };
+    throw error;
+  }
+  if (!stat.isFile()) {
+    throw new Error(`Setup checkpoint path is not a regular file: ${filePath}`);
+  }
+  if (stat.size > MAX_CHECKPOINT_FILE_BYTES) {
+    throw new Error(`Setup checkpoint file exceeds ${MAX_CHECKPOINT_FILE_BYTES} bytes: ${filePath}`);
+  }
+  return {
+    path: filePath,
+    exists: true,
+    data: fs.readFileSync(filePath),
+    mode: stat.mode & 0o777,
+  };
+}
+
+function restoreRegularFile(snapshot, platform = process.platform) {
+  if (!snapshot.exists) {
+    fs.rmSync(snapshot.path, { force: true });
+    return;
+  }
+  writePrivateFileAtomic(snapshot.path, snapshot.data);
+  if (platform !== "win32") fs.chmodSync(snapshot.path, snapshot.mode);
+}
+
+function regularFileChanged(snapshot, platform = process.platform) {
+  let stat;
+  try {
+    stat = fs.lstatSync(snapshot.path);
+  } catch (error) {
+    if (error?.code === "ENOENT") return snapshot.exists;
+    throw error;
+  }
+  if (!snapshot.exists || !stat.isFile()) return true;
+  if (platform !== "win32" && (stat.mode & 0o777) !== snapshot.mode) return true;
+  if (stat.size > MAX_CHECKPOINT_FILE_BYTES) return true;
+  return !fs.readFileSync(snapshot.path).equals(snapshot.data);
+}
+
+function parseBridgeRouteResult(stdout, { expectedActive, requireInstalled = false } = {}) {
+  let result;
+  try {
+    result = JSON.parse(stdout);
+  } catch {
+    throw new Error("Codex bridge route command returned invalid JSON");
+  }
+  if (typeof result?.active !== "boolean") {
+    throw new Error("Codex bridge route command did not report its active state");
+  }
+  if (requireInstalled && typeof result.installed !== "boolean") {
+    throw new Error("Codex bridge route status did not report whether the integration is installed");
+  }
+  if (Array.isArray(result.errors) && result.errors.length > 0) {
+    throw new Error(`Codex bridge route is inconsistent: ${result.errors.join("; ")}`);
+  }
+  if (typeof expectedActive === "boolean" && result.active !== expectedActive) {
+    throw new Error(`Codex bridge route remained ${result.active ? "connected" : "disconnected"}`);
+  }
+  return result;
+}
+
+
+module.exports = {
+  MAX_CAPTURE_BYTES,
+  MAX_RUNTIME_LOG_LINE_CHARS,
+  CORE_SETUP_TIMEOUT_MS,
+  MCP_SETUP_TIMEOUT_MS,
+  UNINSTALL_TIMEOUT_MS,
+  MAX_CHECKPOINT_FILE_BYTES,
+  collect,
+  resolveUserPath,
+  captureRegularFile,
+  restoreRegularFile,
+  regularFileChanged,
+  parseBridgeRouteResult,
+};
