+++ launcher/electron/runtime-setup-transaction.cjs
@@ -1,0 +1,96 @@
+const path = require("node:path");
+const fs = require("node:fs");
+const os = require("node:os");
+const { randomBytes } = require("node:crypto");
+const { spawn } = require("node:child_process");
+const { writePrivateFileAtomic } = require("./atomic-file.cjs");
+const { readJsonFile } = require("./json-file.cjs");
+const { connectorNameForSetup, CURRENT_CONNECTOR_NAME, isLegacyConnectorName, requireCurrentRuntimeConnectorName, validateConnectorName } = require("./connector-identity.cjs");
+const { embeddedRuntimeInvocation, runtimeInvocation } = require("./runtime-command.cjs");
+const { redactText } = require("./logging.cjs");
+const { DETACH_OWNED_CHILD, terminateOwnedProcessTree } = require("./process-tree.cjs");
+const helpers = require("./runtime-helpers.cjs");
+const { MAX_CAPTURE_BYTES, MAX_RUNTIME_LOG_LINE_CHARS, CORE_SETUP_TIMEOUT_MS, MCP_SETUP_TIMEOUT_MS, UNINSTALL_TIMEOUT_MS, MAX_CHECKPOINT_FILE_BYTES, collect, resolveUserPath, captureRegularFile, restoreRegularFile, regularFileChanged, parseBridgeRouteResult } = helpers;
+
+module.exports = {
+  async runDevSetup(name, args, options) {
+    if (this.launcherProfile !== "development") {
+      throw new Error("DEV setup transaction requires the isolated DEV launcher");
+    }
+    return this.runSetup(name, args, {
+      ...options,
+      embedded: true,
+      environment: this.devSetupEnvironment(),
+    });
+  },
+
+  async runSetup(name, args, options) {
+    if (this.currentOperation()) throw new Error(`Another launcher operation is active: ${this.currentOperation()}`);
+    const previousRuntime = this.runtimeConfigSnapshot();
+    const checkpoint = this.captureSetupCheckpoint(previousRuntime);
+    this.lifecycleOperation = name;
+    let setupCommandStarted = false;
+    try {
+      if (previousRuntime.owner === "external") this.supervisor.prepareExternalMigration();
+      else await this.supervisor.stopForSetup();
+      setupCommandStarted = true;
+      const result = await this.run(name, args, options);
+      const runtime = await this.supervisor.startIfConfigured();
+      if (runtime.status !== "ready") {
+        throw new Error(`Setup completed, but the launcher-owned runtime is ${runtime.status}: ${runtime.detail || "not ready"}`);
+      }
+      return result;
+    } catch (error) {
+      const primary = error instanceof Error ? error.message : String(error);
+      const failures = [];
+      let rolledBack = false;
+      let checkpointChanged = false;
+      if (!previousRuntime.configured && setupCommandStarted) {
+        try {
+          rolledBack = await this.rollbackFirstSetup(checkpoint);
+        } catch (caught) {
+          failures.push(
+            `first-time setup rollback failed: ${caught instanceof Error ? caught.message : String(caught)}`,
+          );
+        }
+      }
+      if (previousRuntime.configured && checkpoint) {
+        try {
+          checkpointChanged = this.setupCheckpointChanged(checkpoint);
+        } catch (caught) {
+          checkpointChanged = true;
+          failures.push(
+            `checking the setup checkpoint failed: ${caught instanceof Error ? caught.message : String(caught)}`,
+          );
+        }
+        try {
+          this.restoreSetupCheckpoint(checkpoint);
+        } catch (caught) {
+          failures.push(caught instanceof Error ? caught.message : String(caught));
+        }
+      }
+      let recoveryError;
+      try {
+        await this.restorePreviousRuntime(previousRuntime, name, {
+          repairExternal: previousRuntime.owner === "external" && checkpointChanged,
+        });
+      } catch (caught) {
+        recoveryError = caught;
+      }
+      if (recoveryError) {
+        failures.push(
+          `restoring the previous launcher runtime failed: ${recoveryError instanceof Error ? recoveryError.message : String(recoveryError)}`,
+        );
+      }
+      const message = [
+        primary,
+        ...(rolledBack ? ["incomplete first-time setup was rolled back"] : []),
+        ...failures,
+      ].join("; ");
+      this.publishOperation?.({ name, status: "failed", message });
+      throw new Error(message);
+    } finally {
+      this.lifecycleOperation = null;
+    }
+  },
+};
